requirements.txt
# Market Sensemaking - Production Dependencies
# =============================================

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
starlette>=0.35.0

# Database
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.11
asyncpg>=0.29.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Clients
httpx>=0.26.0
requests>=2.31.0

# WebSocket
websocket-client>=1.7.0
websockets>=12.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Configuration
python-dotenv>=1.0.0

# Visualization (for Streamlit dashboard)
streamlit>=1.30.0
plotly>=5.18.0

# Async Utilities
anyio>=4.2.0

# Date/Time
python-dateutil>=2.8.2

# Cryptography (for evidence hashing)
cryptography>=41.0.0

# JSON handling
orjson>=3.9.0
